refactor: 将 assistant skill 选择工具归集至 agents/utils
这些工具函数仅被 agents 模块使用，从 settings 迁移至 agents/utils 可减少跨模块依赖，提升内聚性
